feat: migrate legend component
