--- src/legend/index.js
@@ -1,3 +1,0 @@
-import Legend from './Legend';
-
-export default Legend;
