+++ src/legend/index.tsx
@@ -1,0 +1,5 @@
+import Legend, { LegendProps, LegendItem } from './Legend';
+
+export type { LegendProps, LegendItem };
+
+export default Legend;
